pyproject.toml
[tool.poetry]
name = "dvtapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apache-beam = {extras = ["gcp"], version = "^2.61.0"}
google-cloud-bigquery = "^3.27.0"
google-cloud-storage = "^2.19.0"
google-auth = "^2.37.0"
poetry-plugin-dotenv = "^2.6.0"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
run_pipeline = "dvtapp.pipeline_8:main"